isFullScreen, etc.
New properties isFullScreen, isVisible, isZoomed and window in class App.

Redefined property isFull as "isFullScreen or isZoomed" in classes App and Window.

Main menu item Full Screen is shown as disabled when the App isFull.

Tested with Python 3.12.2 on macOS 15.3.1 Sequoia and 14.7.3 Sonoma and on AppleSI, Intel and Intel emulation.